requirements.txt
azure-ai-documentintelligence>=1.0.0
pymupdf>=1.24.0
requests>=2.31.0
python-dotenv>=1.0.0
anthropic>=0.40.0
